safari-534.53-branch/LayoutTests/svg/animations/animate-path-nested-transforms-expected.txt

CONSOLE MESSAGE: line 2: ReferenceError: Can't find variable: createSVGTestCase
SVG 1.1 dynamic animation tests

test to determine whether auto-rotate animateMotion path animations pre-muliply or post-multiply animation transform matrix

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
